requirements.txt
asgiref==3.5.2
Django==4.0.6
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
PyJWT==1.7.1
pytz==2022.1
sqlparse==0.4.2
gunicorn==20.1.0
django-shortcuts==1.6
tzdata==2022.6
dj-database-url==1.0.0

